# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hola Municipality Inventory Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List units
      responses:
        "200":
          description: Units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Get unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Deactivate unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # STORES
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH
...
    Unit:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        abbreviation:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    UnitInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        abbreviation:
          type: string
        description:
          type: string
    UnitUpdate:
      type: object
      properties:
        name:
          type: string
        abbreviation:
          type: string
        description:
          type: string
        isActive:
          type: boolean

    # STORE
...